Website Functionality completed.
Added ability to track player and computer score, and update the cookie accordingly.

Added the ability to determine the active user based on URL parameters and reflect the correct user info on the page

Also added logic which automatically navigates the user to index if the username is missing from the url.